Organized the files a bit
