+++ src/lib/mod.rs
@@ -1,0 +1,20 @@
+mod action;
+mod actions;
+mod error;
+mod function;
+mod parser;
+pub mod statics;
+mod types;
+mod utils;
+
+pub use action::{Action, ActionToExpect, ParseAction};
+pub use actions::{Actions, ParseActions};
+pub use error::{ParsingError, ParsingErrorType};
+pub use function::{Function, ParseFunction};
+pub use parser::{CodeLocation, Parser};
+pub use std::fmt::Display;
+pub use types::{ParseType, Type};
+pub use utils::legal_name_char;
+
+#[cfg(test)]
+mod tests;
