+++ src/lib/statics.rs
@@ -1,0 +1,5 @@
+pub static CONST_KEYWORD: &'static str = "const";
+pub static LET_KEYWORD: &'static str = "let";
+pub static RETURN_KEYWORD: &'static str = "return";
+pub static VALID_NAME_CHARS: &'static str =
+  "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890_";
