+++ src/lib/tests/general.rs
@@ -1,0 +1,6 @@
+use super::*;
+
+#[test]
+fn test_empty() {
+  parse_str(r#""#);
+}
