+++ src/lib/utils.rs
@@ -1,0 +1,5 @@
+use super::*;
+
+pub fn legal_name_char(c: char) -> bool {
+  statics::VALID_NAME_CHARS.contains(c)
+}
